feat(hfbcat): Move get_doppler_shifted_freq from ephemeris
This isn't going to go with the rest of the ephemeris
code to the new repository, so let's put is somewhere else.

Weirdly, the original docstring for this function seems to
indicate it was meant to go in `tools`, but since its hardcoded
to work work only on the `HFBCatalog` putting it in `hfbcat`
directly seems most intuitive.

The re-constituted stub in `ephermis` avoids the circular import.

Changes to other files are from `black`.
